+++ lib/src/plugins/versioning.rs
@@ -192,13 +192,13 @@ mod test {
         resource
             .set_propval_string(crate::urls::DESCRIPTION.into(), first_val, &store)
             .unwrap();
-        let first_commit = resource.save_locally(&store).unwrap();
+        let first_commit = resource.save_locally(&store).unwrap().commit_resource;
 
         let second_val = "Hello universe";
         resource
             .set_propval_string(crate::urls::DESCRIPTION.into(), second_val, &store)
             .unwrap();
-        let second_commit = resource.save_locally(&store).unwrap();
+        let second_commit = resource.save_locally(&store).unwrap().commit_resource;
         let commits = get_commits_for_resource(subject, &store).unwrap();
         assert_eq!(commits.len(), 2);